Changes to all db dto be used my models
