+++ flask_backend/models.py
@@ -1,4 +1,23 @@
-class SlippiActionCounts(db.model):
+from app import db
+from flask_wtf import FlaskForm
+from wtforms import FileField, SubmitField
+from wtforms.validators import InputRequired
+
+
+class UploadFileForm(FlaskForm):
+    file = FileField("File", validators=[InputRequired()])
+    submit = SubmitField("Upload File")
+
+class SlippiFile(db.Model):
+    """A slippi action counts"""
+    id = db.Column(db.Integer, primary_key=True)
+    filename = db.Column(db.String(100), nullable=False)
+    connect_code1 = db.Column(db.String(10), nullable=False)
+    connect_code2 = db.Column(db.String(10), nullable=False)
+
+    def __repr__(self):
+        return "SlippiOverall('{}','{}','{}','{}','{}','{}','{}','{}','{}','{}','{}','{}','{}','{}')".format(self.filename, self.connect_code, self.input_counts, self.total_damage, self.kill_count, self.successful_conversions, self.successful_conversion_ratio, self.inputs_per_minute, self.digital_inputs_per_minute, self.openings_per_kill, self.damage_per_opening, self.neutral_win_ratio, self.counter_hit_ratio, self.beneficial_trade_ratio, self.datetime)
+class SlippiActionCounts():
     """A slippi action counts"""
 
     def __init__(self, filename, connect_code, wavedash, waveland, airdodge, dashdance, spotdodge, ledgegrab, roll, lcancel_success_ratio, grab_success, grab_fail, tech_away, tech_in, tech, tech_fail, wall_tech_success_ratio, datetime):
